backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

torch>=2.1.0
torchvision>=0.16.0

transformers==4.41.2
accelerate==0.30.1

pillow==10.3.0
python-multipart==0.0.9

numpy>=1.24.0
